requirements.txt
json
numpy 
os
pillow
replicate
requests
torch
